AI-Code-Reviewer/backend/requirements.txt
fastapi
uvicorn
flake8
black
radon
reportlab
pandas
python-multipart
cpplint
lizard
requests
python-dotenv